Fixed some compilation issues on OS X after reading WinRT changes.

+++ ePub3/utilities/CPUCacheUtils.c
@@ -8,7 +8,7 @@
 
 #include "CPUCacheUtils.h"
 
-int epub_sys_cache_control(int operation, void* start, size_t len)
+int epub_sys_cache_control(int operation, void* start, long len)
 {
     switch ( operation )
     {
